# aula 07/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto de locadora de filmes
  description: API destinada a controle de locadora
  termsOfService: http://swagger.io/terms/
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://
paths:
  /v1/locadora/filme:
    post:
      tags:
        - aluno
      summary: Adiciona um novo Aluno
      description: Adiciona um novo Aluno
      requestBody:
        description: JSON contendo os dados do anovo aluno a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '201':
          description: Successful operation
        '400':
          description: Invalid input
        '500':
          description: Error model
        '500':
          description: Error controller
    get:
      tags:
        - filme
      summary: Listar todos os filmes
      description: Listagem de todos os filmes registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filme'          
           
        '404':
          description: Not Found
  /v1/filme/{id}:
    get:
      tags:
        - filme
      summary: Busca um filme pelo ID
      description: Retorna um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para retornar o filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'          
        '400':
          description: Invalid ID supplied
        '404':
          description: Filme not found
    put:
      tags:
        - filme
      summary: Atualizar um filme pelo ID
      description: Atualizar um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para localizar e atualizar o filme
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar o filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
    delete:
      tags:
        - filme
      summary: Delete filme
      description: Excluir um filme pelo id
      parameters:
        - name: id
          in: path
          description: ID para localizar um filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
  

components:
  schemas:
    Filme:
      required:
        - nome
        - sinopse
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        sinopse:
          type: string
        data_lancamento:
          type: string
        duracao:
          type: string
        valor:
          type: decimal
        avaliacao:
          type: decimal
        capa:
          type: string
    a:
      required:
        - nome
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        celular:
          type: string
        email:
          type: string
  requestBodies:
    Filme:
      description: Objeto que deve ser enviado no Body das requisicoes com os dados do filme
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Filme'
    